Move temperature damage logic to Content.Shared
